Add files via upload
This utility project consists of various microservices. There are EmpService1 and EmpService2 in which we are doing load-balancing using NETFLIX-RIBBON. 
The utility of common configuration using Spring Cloud Config Server is also shown in the same .
Maven build ---->Import in your STS----> Update the Maven project and you are good to go.